refactor(project-structure): rename packages to plural form
